GNNDrugDiscovery/gnn_midterm_starter_v2/requirements.txt
torch
torch-geometric
rdkit-pypi
numpy
pandas
scikit-learn
tqdm
matplotlib
PyYAML
